% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/buildMapping.R
\docType{methods}
\name{buildMapping}
\alias{buildMapping}
\alias{buildMapping}
\alias{buildMapping,GeneGOMapping-method}
\title{Method buildMapping.}
\usage{
\S4method{buildMapping}{GeneGOMapping}(object, header = FALSE)
}
\arguments{
\item{object}{the GeneGOMapping object}

\item{header}{input mapping file has a header or not}
}
\value{
a GeneGOMapping with list of GO to Gene mapping
}
\description{
Method buildMapping.
}
\examples{
#relationship between the gene and GO terms
#sep used in the input file by default is tab
inputFile <- paste0(system.file(package = "GSminer"), "/extdata/TAIR.GO")
miner1 <- GSminer(inputFile = inputFile, sep = "\\t")
buildMapping(miner1@ggmapping)
}
\author{
Li Zhaohong && Wu Zefeng
}
