% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\docType{methods}
\name{filterNeg,GSminer-method}
\alias{filterNeg,GSminer-method}
\title{Filter neagtive paris by cytoplasm and nucleus cell compartments}
\usage{
\S4method{filterNeg}{GSminer}(object, negFilename = "negativeGene.txt",
  FNegFile = "negativeGene.filtered.txt", sep = "\\t")
}
\arguments{
\item{object}{the GSminer object}

\item{negFilename}{the negative gene file name used for filtering}

\item{FNegFile}{the output file name which is filtered by cytoplasm and nucleus cell compartments}

\item{sep}{the delimiter of the file}
}
\value{
neagtive paris file
}
\description{
Filter neagtive paris by cytoplasm and nucleus cell compartments
}
\examples{

\dontrun{
inputFile <- paste0(system.file(package = "GSminer"), "/extdata/TAIR.GO")
library(GSminer)
test <- GSminer(inputFile = inputFile)
filterNeg(object, negFilename = "negFilename", FNegFile = "FNegFile")
\\}
}
