% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/buildMapping.R
\docType{methods}
\name{gene2BP}
\alias{gene2BP}
\alias{gene2BP}
\alias{gene2BP,GeneGOMapping-method}
\title{Method gene2BP.}
\usage{
\S4method{gene2BP}{GeneGOMapping}(object)
}
\arguments{
\item{object}{the GeneGOMapping object}
}
\value{
a list of GO to Gene mapping
}
\description{
Method gene2BP.

gene for GO Biological Process (BP) terms
}
\examples{

inputFile <- paste0(system.file(package = "GSminer"), "/extdata/TAIR.GO")
#inputFile the name of input file which contain the mapping 
#relationship between the gene and GO terms
#sep used in the input file by default is tab
#for an example how to use:
miner <- GSminer(inputFile = inputFile, sep = "\\t")
#gene2BP(miner@ggmapping)
}
\author{
Li Zhaohong && Wu Zefeng
}
