% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{GeneGOMapping-class}
\alias{GeneGOMapping-class}
\alias{GeneGOMapping}
\title{Class "GeneGOMapping"
A S4 class for mapping between gene and GO terms
This class is the solt of the GSminer which used to select genes.}
\description{
Class "GeneGOMapping"
A S4 class for mapping between gene and GO terms
This class is the solt of the GSminer which used to select genes.
}
\section{Slots}{

\describe{
\item{\code{inputFile}}{the name of input file which contain the mapping 
relationship between the gene and GO terms}

\item{\code{sep}}{used in the input file by default is tab}

\item{\code{mapping}}{is a private env}
}}

\examples{

inputFile <- paste0(system.file(package = "GSminer"), "/extdata/TAIR.GO")
ggmap <- GeneGOMapping(inputFile = inputFile, sep = "\\t")
###	
ggmap <- methods::new("GeneGOMapping", inputFile = inputFile, 
                        sep = "\\t")	

}
\seealso{
\code{\linkS4class{GSminer}}
}
\author{
Li Zhaohong && Wu Zefeng
}
\keyword{classes}
