% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSim.R
\name{geneSim}
\alias{geneSim}
\title{The similarity between genes(modified the dDAGgeneSim in dnet)}
\usage{
geneSim(g, genes = NULL, method.gene = c("BM.average", "BM.max",
  "BM.complete", "average", "max"), method.term = c("Resnik", "Lin",
  "Schlicker", "Jiang", "Pesquita"), force = TRUE, fast = TRUE,
  parallel = TRUE, multicores = NULL, verbose = TRUE)
}
\arguments{
\item{g}{an object of class "igraph"}

\item{genes}{the genes between which pair-wise semantic similarity is calculated}

\item{method.gene}{the method used for how to derive semantic similarity
between genes from semantic similarity between terms}

\item{method.term}{the method used to measure semantic similarity between terms}

\item{force}{logical}

\item{fast}{logical to indicate whether a vectorised fast computation is used}

\item{parallel}{logical to indicate whether parallel computation}

\item{multicores}{parallel computation}

\item{verbose}{logical}
}
\value{
a sparse matrix of similarity between input genes
}
\description{
The similarity between genes(modified the dDAGgeneSim in dnet)
}
\examples{
#load HPPA as igraph object
HPPA <-dRDataLoader(RData='ig.HPPA')
g <- ig.HPPA
#load human genes annotated by HPPA
Hs.egHPPA <- dRDataLoader(RData='org.Hs.egHPPA')
dag <- dDAGannotate(g, annotations=org.Hs.egHPPA,
path.mode="all_paths", verbose=TRUE)
allgenes <- unique(unlist(V(dag)$annotations))
genes <- sample(allgenes,5)
#just for an example:
#geneSim(g = dag, genes = genes, method.gene="BM.average", 
#        method.term = "Resnik", parallel = TRUE, 
#        multicores = 2, verbose = FALSE)


}
