% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getGOSim}
\alias{getGOSim}
\title{getGOSim}
\usage{
getGOSim(g, selectedGO, method = c("Resnik", "Lin", "Schlicker", "Jiang",
  "Pesquita"), multicores = NULL)
}
\arguments{
\item{g}{igraph object}

\item{selectedGO}{list which name is selected GO term}

\item{method}{method for computing the similarity}

\item{multicores}{an integer to specify how many cores will be registered as the multicore parallel backend to the 'foreach' package.}
}
\value{
similarity matrix
}
\description{
Compute the GO term similarity
}
\examples{

library(igraph)
#just for an example:
#HPPA <-dRDataLoader(RData='ig.HPPA')
#g <- ig.HPPA
#getGOSim(g, selectedGO, method= "Resnik", multicores = 3)

}
\author{
Li Zhaohong && Wu Zefeng
}
