% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getThreshold.R
\docType{methods}
\name{getThreshold}
\alias{getThreshold}
\alias{getThreshold}
\alias{getThreshold,GSminer-method}
\title{getThreshold method generics}
\usage{
\S4method{getThreshold}{GSminer}(object, u = 0.5, l = 0.25)
}
\arguments{
\item{object}{the GSminer object}

\item{u}{the upper bound, 1/2 by default}

\item{l}{the lower bound, 1/4 by default}
}
\value{
a list of upper bound and lower bound
}
\description{
getThreshold method generics

select the GO term size
}
\examples{

inputFile <- paste0(system.file(package = "GSminer"), "/extdata/TAIR.GO")
miner <- GSminer(inputFile = inputFile, sep = "\\t")
# just for an example:
#getThreshold(miner, u = 0.5, l = 0.25)
}
\author{
Li Zhaohong && Wu Zefeng
}
