% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{GO2igraph}
\alias{GO2igraph}
\title{convert to graph format}
\usage{
GO2igraph(mapping = GOBPPARENTS)
}
\arguments{
\item{mapping}{the mapping which convert to igraph object}
}
\value{
igraph object
}
\description{
Format conversion
}
\examples{
library(GO.db)
GO2igraph(mapping = GOBPPARENTS)

}
\author{
Li Zhaohong && Wu Zefeng
}
