% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{GOCluster}
\alias{GOCluster}
\title{GOCluster}
\usage{
GOCluster(sim, selectedGO, seed)
}
\arguments{
\item{sim}{similarity matrix}

\item{selectedGO}{list which name is selected GO term}

\item{seed}{for reproducibility,}
}
\value{
final Selected GO terms
}
\description{
Clustering by apcluster to get the optimal cluster number and get the final go terms
}
\examples{
\dontrun{
library(apcluster)
GOCluster(sim, selectedGO, seed = 100)
}

}
\author{
Li Zhaohong && Wu Zefeng
}
