% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{GSminer-class}
\alias{GSminer-class}
\alias{GSminer}
\alias{GSminerClass}
\title{Class "GSminer"}
\arguments{
\item{inputFile}{input file name(with absolute path)}

\item{sep}{input file separator by default is tab}
}
\description{
GSminer has three slots:
1. ggmapping object for gene to GO terms mapping
2. BPoffsprings a list for BP GO terms to gene mapping
3. CCoffsprings a list for CC GO terms to gene mapping
Note: the BPoffspring and CCoffspring mimic the behavior of pass by const reference in C++
}
\section{Slots}{

\describe{
\item{\code{ggmapping}}{which is the instance of GeneGOMapping for gene to GO terms mapping}

\item{\code{mapping}}{is a private env}
}}

\examples{

inputFile <- paste0(system.file(package = "GSminer"), "/extdata/TAIR.GO")
miner <- GSminer(inputFile = inputFile, sep = "\\t")
ggmap <- methods::new("GeneGOMapping", inputFile = inputFile, sep = "\\t")
miner@ggmapping <- ggmap
loadBPoffsprings(miner)

}
\seealso{
\code{\linkS4class{GeneGOMapping}}
}
\author{
Li Zhaohong && Wu Zefeng
}
\keyword{classes}
