% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/loadBPoffsprings.R
\docType{methods}
\name{loadBPoffsprings}
\alias{loadBPoffsprings}
\alias{loadBPoffsprings}
\alias{loadBPoffsprings,GSminer-method}
\title{loadBPoffsprings method generics}
\usage{
\S4method{loadBPoffsprings}{GSminer}(object)
}
\arguments{
\item{object}{the GeneGOMapping object}
}
\value{
a list of BP GO to Gene mapping
}
\description{
loadBPoffsprings method generics

load Biological Process GO offsprings for genes
}
\examples{
inputFile <- paste0(system.file(package = "GSminer"), "/extdata/TAIR.GO")
test <- GSminer(inputFile = inputFile)
#for an example:
#loadBPoffsprings(test)
}
\author{
Li Zhaohong && Wu Zefeng
}
