% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/loadBPoffsprings.R
\docType{methods}
\name{loadCCoffsprings}
\alias{loadCCoffsprings}
\alias{loadCCoffsprings}
\alias{loadCCoffsprings,GSminer-method}
\title{loadCCoffsprings method}
\usage{
\S4method{loadCCoffsprings}{GSminer}(object)
}
\arguments{
\item{object}{the GeneGOMapping object}
}
\value{
a list of CC GO to Gene mapping
}
\description{
loadCCoffsprings method

load cytoplasm and nucleus cell compartments GO terms for genes
}
\examples{
inputFile <- paste0(system.file(package = "GSminer"), "/extdata/TAIR.GO")
test <- GSminer(inputFile = inputFile)
#just like loadBPoffsprings
#loadCCoffsprings(test)
}
\author{
Li Zhaohong && Wu Zefeng
}
