% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\name{MappingFileValid}
\alias{MappingFileValid}
\title{Validation of GeneGOMapping object}
\usage{
MappingFileValid(object)
}
\arguments{
\item{object}{instance of GeneGOMapping for gene to GO terms mapping}
}
\value{
TRUE for validation
}
\description{
Validation of GeneGOMapping object
}
