% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/output.R
\docType{methods}
\name{outputPosNeg}
\alias{outputPosNeg}
\alias{outputPosNeg}
\alias{outputPosNeg,GSminer-method}
\title{outputPosNeg method generics}
\usage{
\S4method{outputPosNeg}{GSminer}(object, u = 0.5, l = 0.25,
  method.term = c("Resnik", "Lin", "Schlicker", "Jiang", "Pesquita"),
  method.gene = c("BM.average", "BM.max", "BM.complete", "average", "max"),
  multicores = NULL, verbose = TRUE, seed = NA,
  posFilename = "positiveGene.txt", negFilename = "negativeGene.txt",
  sep = "\\t", filterByCC = TRUE, FNegFile = "negativeGene.filtered.txt")
}
\arguments{
\item{object}{the GSminer object}

\item{u}{the upper bound, 1/2 by default}

\item{l}{the lower bound, 1/4 by default}

\item{method.term}{method for computing the similarity}

\item{method.gene}{the method used for how to derive semantic similarity between genes from semantic similarity between terms.}

\item{multicores}{an integer to specify how many cores will be registered as the multicore parallel backend to the 'foreach' package.}

\item{verbose}{logical to indicate whether the messages will be displayed in the screen.}

\item{seed}{for reproducibility, the seed of the random number generator can be set to a fixed value before adding noise.}

\item{posFilename}{postive gene output file name}

\item{negFilename}{negative gene output file name without filtered by CC GO terms}

\item{sep}{the delimiter of the file}

\item{filterByCC}{weather filtered by CC GO terms by default is TRUE}

\item{FNegFile}{negative gene output file name filtered by CC GO terms}
}
\value{
a vector which name is selected GO term
}
\description{
outputPosNeg method generics

Output negative gene pairs
}
\examples{

inputFile <- paste0(system.file(package = "GSminer"), "/extdata/TAIR.GO")
miner <- GSminer(inputFile = inputFile, sep = "\\t")
#outputPosNeg(miner, u = 0.1, l = 0.2, method.term = "Resnik", 
#             method.gene = "BM.average", multicores = 2, 
#             verbose = FALSE, posFilename = "pos.txt", 
#            filterByCC = FALSE, FNegFile = "neg.text")

}
\author{
Lizhaohong && Wu Zefeng
}
