% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removedup.R
\name{removeDup}
\alias{removeDup}
\title{Remove duolicated rows}
\usage{
removeDup(data)
}
\arguments{
\item{data}{data frame}
}
\value{
an unique data frame
}
\description{
Remove duolicated rows
}
\examples{
# GPSimNeg a data frame
test <- data.frame("a" = c("ab", "cd", "ef"), "b" = c("cd", "ab", "ef")) 
test <- removeDup(test)

}
