% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getThreshold.R
\docType{methods}
\name{selectGO}
\alias{selectGO}
\alias{selectGO}
\alias{selectGO,GSminer-method}
\title{selectGO method generics}
\usage{
\S4method{selectGO}{GSminer}(object, u = 0.5, l = 0.25,
  method = c("Resnik", "Lin", "Schlicker", "Jiang", "Pesquita"),
  multicores = NULL, seed = NA)
}
\arguments{
\item{object}{the GSminer object}

\item{u}{the upper bound, 1/2 by default}

\item{l}{the lower bound, 1/4 by default}

\item{method}{method for computing the similarity}

\item{multicores}{an integer to specify how many cores will be registered as the multicore parallel backend to the 'foreach' package.}

\item{seed}{for reproducibility, the seed of the random number generator can be set to a fixed value before adding noise.}
}
\value{
a vector which name is selected GO term
}
\description{
selectGO method generics

select the GO term based on the threshold
}
\examples{

inputFile <- paste0(system.file(package = "GSminer"), "/extdata/TAIR.GO")
miner <- GSminer(inputFile = inputFile, sep = "\\t")
#just for an example
#selectGO(miner,u = 0.5, l = 0.25, method= "Resnik", multicores = 2)

}
\author{
Li Zhaohong && Wu Zefeng
}
