% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSminer-package.R
\docType{data}
\name{TAIR.GO}
\alias{TAIR.GO}
\alias{mydata}
\title{Datasets
the mapping between Arabidopsis thaliana gene and GO term}
\format{a data frame}
\usage{
mydata
}
\description{
the mapping between Arabidopsis thaliana gene and GO term
}
\keyword{datasets}
